refac: debate hook
