+++ book-talk-fe/src/hooks/usePresentation.tsx
@@ -8,6 +8,7 @@ import {
 import {useCallback, useMemo, useRef, useState} from 'react';
 import {useJsonPatch} from "./useJsonPatch.tsx";
 
+/** 발표 페이지를 저장 및 관리 */
 export const usePresentation = (
     presentationId?: string,
 ) => {
